fix: update exitType uint8 -> uint256 in the twg interface

+++ src/interfaces/ITriggerableWithdrawalsGateway.sol
@@ -31,6 +31,6 @@ interface ITriggerableWithdrawalsGateway {
     function triggerFullWithdrawals(
         ValidatorData[] calldata triggerableExitsData,
         address refundRecipient,
-        uint8 exitType
+        uint256 exitType
     ) external payable;
 }
